fix overflow issue (Comment: #12344)
git-svn-id: https://ut.svn.beanstalkapp.com/ngfw/work/src@40599 8ebf8b4e-06f3-0310-961f-89f74618be54

+++ uvm/impl/com/untangle/uvm/HostTableImpl.java
@@ -219,7 +219,7 @@ public class HostTableImpl implements HostTable
         entry.setQuotaSize( quotaBytes );
         entry.setQuotaRemaining( quotaBytes );
         entry.setQuotaIssueTime( now );
-        entry.setQuotaExpirationTime( now + (time_sec*1000) );
+        entry.setQuotaExpirationTime( now + (((long)time_sec)*1000L) );
 
         /**
          * Call listeners
